Move config and run blocks of auth module into their own files
