+++ client/src/auth/run.ts
@@ -1,0 +1,20 @@
+import { AuthService } from './auth.service';
+
+/**
+ * Inject auth service into run block.
+ */
+run.$inject = [
+  'AuthService'
+];
+
+/**
+ * Initialize auth service.
+ */
+function run(
+  AuthService: AuthService
+): void {
+  console.log('Initializing auth service.');
+  AuthService.syncWithAuth0();
+}
+
+export { run }
